feat: clicks mapper go
* feat: add clicks model
* feat: add register click
* feat: add click service
* refactor: change func name
* refactor: change id type
* feat: setup click analytics

+++ go-redirector/internal/model/url_model.go
@@ -2,7 +2,7 @@ package model
 
 type (
 	UrlModel struct {
-		ID            int    `db:"id"`
+		ID            int64  `db:"id"`
 		Url           string `db:"url"`
 		CustomSlug    string `db:"custom_slug"`
 		ClickCount    int    `db:"click_count"`
